==> 6_comparison.ipynb <==
|      | Precision | Recall | Accuracy |
|------|-----------|--------|----------|
| TF-IDF | 100%     | 83.1%  | 97.8%    |
| word2vec | 60%     | 22.3%  | 87.7%    |
| doc2vec | 86.8%     | 44.6%  | 91.7%    |
| RNN | 98.7%     | 96.9%  | 93.3%    |

These models are not deterministic.

- TF-IDF performs really really well! When the model says a text message is spam, it actually is spam 100% of the time. When a text message actually is spam, it identifies it as such 83.1% of the time. 16.9% of real spam would make it through. When classifying spam or ham, the model was correct 97.3% of the time.

Same interpretation for the others.

- word2vec, due to the averaging we did, is the worst model by far.
- RNN was outstanding but also really close to the baseline.

This is an importante case study, also when considering the precision-recall calibration. In a problem like fraud detection, you need to optimize for recall since you want to really capture fraud without missing any. In a situation of spam filtering you should optimize for precision: in other words, i can handle the model allowing some spam into my inbox but if it classifies a real message as spam i won't be happy!

So in this case, TF-IDF could be the way to go, even over the RNN.

# Key takeaways of these techniques

### TF-IDF
- Simple
- no context
- very sparse vectors
- good baseline

### Word2Vec
- shallow NN that creates word vectors that can be averaged to create a document level representation
- this averaging causes loss of info
- more sophisticated
- dense vectors
- many more

### Doc2vec
- document level representation vectors
- slower but can be powerful
- better than word2vec in sentence level representation
- smaller dense 
- considers context as word2vec with the same window (previous and next n-grams) approach

### RNN
- perfect for sequential data like sentences
- dense vectors are created within the model
- powerful also with limited data
